added Damper test suite
