+++ packages/model-viewer/src/test/three-components/SmoothControls-spec.ts
@@ -15,7 +15,6 @@
 
 import {PerspectiveCamera, Vector3} from 'three';
 
-import {Damper} from '../../three-components/Damper.js';
 import {ChangeSource, DEFAULT_OPTIONS, KeyCode, SmoothControls} from '../../three-components/SmoothControls.js';
 import {dispatchSyntheticEvent, waitForEvent} from '../helpers.js';
 
@@ -36,26 +35,6 @@ const DEFAULT_INTERACTION_CHANGE_SOURCE = 'none';
  */
 export const settleControls = (controls: SmoothControls) =>
     controls.update(performance.now(), FIFTY_FRAME_DELTA);
-
-suite('Damper', () => {
-  let damper: Damper;
-  const initial = 5;
-  const goal = 2;
-
-  setup(() => {
-    damper = new Damper();
-  });
-
-  test('converges to goal with large time step without overshoot', () => {
-    const final = damper.update(initial, goal, FIFTY_FRAME_DELTA, initial);
-    expect(final).to.be.eql(goal);
-  });
-
-  test('stays at initial value for negative time step', () => {
-    const final = damper.update(initial, goal, -1 * FIFTY_FRAME_DELTA, initial);
-    expect(final).to.be.eql(initial);
-  });
-});
 
 suite('SmoothControls', () => {
   let controls: SmoothControls;
